big rewrite for a better integartion with API
