+++ src/lang/en.json
@@ -9,7 +9,7 @@
         "Import": "Import",
         "Close": "Close"
       }
-    },  
+    },
     "UI": {
       "Chat": { "Narrator": "Narrator" },
       "Title": {
