+++ src/scripts/TheatreActorConfig.js
@@ -20,6 +20,7 @@
 
 import { KHelpers } from "./KHelpers.js";
 import { Theatre } from "./Theatre.js";
+import { debug } from "./lib/lib.js";
 
 /**
  * ============================================================
@@ -140,7 +141,7 @@ export class TheatreActorConfig extends FormApplication {
         let labelPath = mch + ".label";
         let cflagPath = mch + ".custom";
         let namePath = mch + ".name";
-        if (Theatre.DEBUG) console.log("found %s", k, mch, cflagPath, namePath);
+        debug("found %s", k, mch, cflagPath, namePath);
         // if label is both the formData as well as the object, reject the submission
         let emoteProp = getProperty(this.object, mch);
         let labelProp = null;
@@ -218,7 +219,7 @@ export class TheatreActorConfig extends FormApplication {
     formData["_id"] = this.object._id;
 
     // if our baseinsert value was updated..
-    if (Theatre.DEBUG) console.log(formData);
+    debug(formData);
     let insertDirty = false;
     let baseInsert = formData["flags.theatre.baseinsert"];
     let optAlign = formData["flags.theatre.optalign"];
@@ -234,7 +235,7 @@ export class TheatreActorConfig extends FormApplication {
     let cImg = Theatre.instance.getTheatreCoverPortrait();
 
     if (baseInsert != this.object.flags.theatre.baseinsert) {
-      if (Theatre.DEBUG) console.log("baseinsert changed!");
+      debug("baseinsert changed!");
       insertDirty = true;
       newBaseInsert = baseInsert == "" ? (this.object.img ? this.object.img : "icons/mystery-man.png") : baseInsert;
       if (navItem) {
@@ -243,13 +244,13 @@ export class TheatreActorConfig extends FormApplication {
       }
     }
     if (optAlign != this.object.flags.theatre.optalign) {
-      if (Theatre.DEBUG) console.log("optalign changed!");
+      debug("optalign changed!");
       insertDirty = true;
       newAlign = optAlign == "" ? "top" : optAlign;
       if (navItem) navItem.setAttribute("optalign", newAlign);
     }
     if (name != this.object.flags.theatre.name) {
-      if (Theatre.DEBUG) console.log("name changed!");
+      debug("name changed!");
       insertDirty = true;
       newName = name == "" ? this.object.name : name;
       if (navItem) {
@@ -262,7 +263,7 @@ export class TheatreActorConfig extends FormApplication {
     // Verify custom emotes if we have any
     let resForms = this._verifyCustomEmotes(formData);
     if (!resForms) return;
-    if (Theatre.DEBUG) console.log("Form data AFTER verification: ", resForms);
+    debug("Form data AFTER verification: ", resForms);
     let revisedFormData = resForms.revisedFormData;
     let emoteFormData = resForms.emoteFormData;
 
@@ -320,13 +321,13 @@ export class TheatreActorConfig extends FormApplication {
     }
     // send the emote parent in bulk to get rid of unwanted children
     revisedFormData["flags.theatre.emotes"] = nEmotes;
-    if (Theatre.DEBUG) console.log("Final Push Config update:", revisedFormData);
+    debug("Final Push Config update:", revisedFormData);
 
     this.object.update(revisedFormData).then((response) => {
       // perform texture updates if needed
       if (imgSrcs.length > 0) {
         // we know the active emote, thus all we need is the new source image
-        if (Theatre.DEBUG) console.log("sending imgSrcs for replaceAllTextures", imgSrcs);
+        debug("sending imgSrcs for replaceAllTextures", imgSrcs);
         Theatre.instance._AddAllTextureResources(
           imgSrcs,
           theatreId,
@@ -334,10 +335,10 @@ export class TheatreActorConfig extends FormApplication {
           newSrcImg,
           (loader, resources) => {
             // if our emote is active and we're replacing the emote texture, or base is active, and we're replacing the base texture
-            if (Theatre.DEBUG) console.log("texture additions complete! ", newSrcImg, insertEmote);
+            debug("texture additions complete! ", newSrcImg, insertEmote);
 
             if (app && container && newSrcImg) {
-              if (Theatre.DEBUG) console.log("RE-RENDERING with NEW texture resource %s ", newSrcImg);
+              debug("RE-RENDERING with NEW texture resource %s ", newSrcImg);
 
               let resName = "icons/myster-man.png";
               if (
@@ -377,7 +378,7 @@ export class TheatreActorConfig extends FormApplication {
 
       // if the insert is dirty, clear and setup
       if (insertDirty && insert) {
-        if (Theatre.DEBUG) console.log("Insert is dirty, re-render it!");
+        debug("Insert is dirty, re-render it!");
         let resName = "icons/myster-man.png";
         if (
           insert.emote &&
@@ -428,7 +429,7 @@ export class TheatreActorConfig extends FormApplication {
    * @private
    */
   _onAddEmoteLine(ev) {
-    if (Theatre.DEBUG) console.log("Add Emote Pressed!");
+    debug("Add Emote Pressed!");
     //ui.notifications.info(game.i18n.localize("Theatre.NotYet"));
 
     // We need to get a custom emote name for storage purposes, this is a running index from
@@ -669,7 +670,7 @@ export class TheatreActorConfig extends FormApplication {
    * @private
    */
   _onUndoDockDelete(ev) {
-    if (Theatre.DEBUG) console.log("undo delete!");
+    debug("undo delete!");
     ev.stopPropagation();
     ev.currentTarget.removeAttribute("todelete");
     ev.currentTarget.style.left = "0";
@@ -700,7 +701,7 @@ export class TheatreActorConfig extends FormApplication {
    * @private
    */
   _onEditEmoteLine(ev) {
-    if (Theatre.DEBUG) console.log("Emote config pressed for %s!", ev.currentTarget.getAttribute("name"));
+    debug("Emote config pressed for %s!", ev.currentTarget.getAttribute("name"));
     ui.notifications.info(game.i18n.localize("Theatre.NotYet"));
   }
 }
