+++ wiki/macros/add-selected_tokens_to_the_stage.js
@@ -1,0 +1,3 @@
+for (const tkn of canvas.tokens.controlled) {
+    Theatre.addToNavBar(tkn.actor);
+}
