+++ wiki/macros/add_owned_tokens_to_the_stage.js
@@ -1,0 +1,6 @@
+const ownedActors = game.actors.filter(a => a.permission === 3);
+const ownedTokens = ownedActors.map(a => a.getActiveTokens());
+
+for (const tokenArray of ownedTokens) {
+    tokenArray.forEach(t => Theatre.addToNavBar(t.actor));
+}
